chat.seoeunjin.com/backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.1  # Midm 모델 사용 시 필요
psycopg2-binary>=2.9.0
pgvector>=0.2.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
# Midm 모델 사용 시 필요 (CPU 버전)
# torch는 배포 스크립트에서 CPU 전용 버전으로 별도 설치
transformers>=4.39.0
accelerate>=0.20.0
python-dotenv>=1.0.0

